@@ -14,15 +14,15 @@
         "issues": "https://github.com/mhrabovcin/zuora-rest/issues"
     },
     "require": {
-        "php": ">=5.3.0"
+        "php": "^7.1"
     },
     "require-dev": {
         "pdepend/pdepend": "~1.0",
-        "phploc/phploc": "~2.0",
         "phpmd/phpmd": "~1.0",
-        "phpunit/phpunit": "~3.0",
         "satooshi/php-coveralls": "*",
-        "sebastian/phpcpd": "~2.0"
+        "sebastian/phpcpd": "~2.0",
+        "phpunit/phpunit": "^7.5",
+        "phploc/phploc": "^4.0"
     },
     "autoload": {
         "psr-0": {
