@@ -2,12 +2,12 @@
 
 namespace Zuora\Test;
 
-
 use Zuora\Client;
-use Zuora\Environment;
+use Zuora\Http\RequestInterface;
 use Zuora\Http\Response;
 
-class ResponseTest extends \Zuora\Test\Base {
+class ResponseTest extends Base
+{
 
     /**
      * Tests Response pagination and class mapping
@@ -16,38 +16,39 @@ class ResponseTest extends \Zuora\Test\Base {
     {
         $environment = $this->getEnvironment();
 
-        $response_data = array(
-           'success' => true,
-           'creditCards' => array(
-              array('id' => md5(rand())),
-           ),
-           'nextPage' => $environment->getUrl('accounts') . '?page=2',
-        );
+        $response_data = [
+            'success' => true,
+            'creditCards' => [
+                ['id' => md5(rand())],
+            ],
+            'nextPage' => $environment->getUrl('accounts') . '?page=2',
+        ];
 
         $http_response = new Response();
         $http_response->setData($response_data)
             ->setCode(200)
-            ->setHeaders(array());
+            ->setHeaders([]);
 
         $next_response = new Response();
         $next_response->setCode(200)
-            ->setHeaders(array())
-            ->setData(array('nextPage' => $environment->getUrl('accounts') . '?page=3') + $response_data);
+            ->setHeaders([])
+            ->setData(['nextPage' => $environment->getUrl('accounts') . '?page=3'] + $response_data);
 
 
-        $request = $this->getMock('\Zuora\Http\RequestInterface');
+        $request = $this->createMock(RequestInterface::class);
         $request->expects($this->once())
-           ->method('call')
-           ->with($this->equalTo($environment->getUrl('accounts')),
-                  $this->equalTo('GET'),
-                  $this->equalTo(array('page' => 2)),
-                  $this->anything(),
-                  $this->equalTo(array(
-                     'apiAccessKeyId' => $environment->getUsername(),
-                     'apiSecretAccessKey' => $environment->getPassword(),
-                  ))
-           )
-           ->will($this->returnValue($next_response));
+            ->method('call')
+            ->with(
+                $this->equalTo($environment->getUrl('accounts')),
+                $this->equalTo('GET'),
+                $this->equalTo(['page' => 2]),
+                $this->anything(),
+                $this->equalTo([
+                    'apiAccessKeyId' => $environment->getUsername(),
+                    'apiSecretAccessKey' => $environment->getPassword(),
+                ])
+            )
+            ->will($this->returnValue($next_response));
 
         $client = new Client($environment, $request);
 
@@ -59,8 +60,7 @@ class ResponseTest extends \Zuora\Test\Base {
         $cards = $response->map('creditCards', '\Zuora\Object\CreditCard');
         $this->assertInternalType('array', $cards);
 
-        $response = new \Zuora\Response($http_response->setData(array()), $client);
+        $response = new \Zuora\Response($http_response->setData([]), $client);
         $this->assertNull($response->nextPage());
     }
-
-} 
+}
