@@ -1,11 +1,9 @@
 <?php
 
-
 namespace Zuora\Object;
 
-
-class Usage extends Object {
-
+class Usage extends ZuoraObject
+{
     /**
      * @return mixed
      */
@@ -101,5 +99,4 @@ class Usage extends Object {
     {
         return $this->unitOfMeasure;
     }
-
-} 
+}
